requirements.txt
Django==1.6.1
Pillow==2.3.0
SPARQLWrapper==1.5.2
SQLAlchemy==0.8.1
South==0.7.6
Unidecode==0.04.14
distribute==0.6.34
django-extensions==1.1.1
django-haystack==2.1.0
django-mptt==0.5.-dev
django-pipeline==1.3.20
futures==2.1.6
html5lib==1.0b1
isodate==0.4.9
lxml==3.2.1
psycopg2==2.5
pyelasticsearch==0.6.1
pyparsing==1.5.7
python-dateutil==2.1
rdflib==4.0.1
rdflib-sqlalchemy==0.2.dev
requests==2.2.1
simplejson==3.3.2
six==1.3.0
wikitools==1.1.1
wsgiref==0.1.2

# requirement for server. Please select the ones that applies to ypur environment
#python-memcached
#pylibmc
#uWSGI
#pytz